+++ CHANGELOG.md
@@ -6,6 +6,8 @@ _The format is based on [Keep a Changelog](http://keepachangelog.com/) and this 
 
 ## Unreleased
 
+- Remove deprecated functionality (object-form `.attrs({})`)
+
 ## [v4.3.1] - 2019-06-06
 
 - Revert #2586; breaks rehydration in dev for certain runtimes like next.js